improved update handling, server requested restart
Handle onUpdateAvailable correctly by doing a restart. The restart is
delayed to between 4 and 5 at night.

Check every 5 minutes (after initial delay of 1 minute) for restart
signal from the server and restart when correct signal is received.

Store all the OS info in variables

When app is closed stop the possibly scheduled restart and the check for
restart