increased version, set offline_enabled to true
The documentation states:
As of Chrome 35, apps are assumed to be offline enabled and the default value of "offline_enabled" is true unless "webview" permission is requested. In this case, network connectivity is assumed to be required and "offline_enabled" defaults to false.

The "offline_enabled" value is also used to determine whether a network connectivity check will be performed when launching an app in Chrome OS kiosk mode. A network connectivity check will be performed when apps are not offline enabled, and app launching put on hold until the device obtains connectivity to the Internet.

The last sentence was interpreted as meaning that staring the kiosk app would only
be delayed by the fact that there is no ntwork connection when starting
up the Chrome device.
However, it turns out that in such a case the kiosk app is not started,
an error text is shown and the kiosk app is not started after that. Even
when a network connenction is subsequently established and the device is
able to report its state as running to the Google Device Management and
a screen shot of the error message can be made using thsi service.

In an attempt to prevent this situation from which the device cannot
recover by itself we set offline_enabled to true and rely on the network
check that is already present in the main.html and main.js.